[clear] fake data
