+++ src/components/MessageItem/index.ts
@@ -1,2 +1,1 @@
 export * from './MessageItem';
-export * from './data';
