+++ src/utils/index.ts
@@ -1,4 +1,4 @@
 export * from './pages';
 export * from './helpers';
-export * from './faker';
+export * from './contentBuildes';
 export * from './commonHandlers';
